requirements.txt
streamlit
fastapi
pydantic
python-dotenv
qdrant-client
google-generativeai
